feat: new functions
